\name{A}
\alias{A}

\docType{data}

\title{
Adjacency matrix for the infant mortality data.
}

\description{
A contains the adjacency matrix for the infant mortality data analyzed in (Hughes and Haran, 2013).
}

\usage{data(A)}

\source{
The data were obtained from the 2008 Area Resource File (ARF), a county-level database maintained by the Bureau of Health Professions, Health Resources and Services Administration, US Department of Health and Human Services.
}

\examples{
data(A)
}

\keyword{datasets}
