\name{infant}
\alias{infant}

\docType{data}

\title{
Infant mortality data.
}

\description{
infant contains the infant mortality data analyzed in (Hughes and Haran, 2013).
}

\usage{data(infant)}

\source{
The data were obtained from the 2008 Area Resource File (ARF), a county-level database maintained by the Bureau of Health Professions, Health Resources and Services Administration, US Department of Health and Human Services.
}

\examples{
data(infant)
}

\keyword{datasets}
