% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.msglmm.R
\name{sparse.msglmm}
\alias{sparse.msglmm}
\title{Fit a sparse multivariate SGLMM.}
\usage{
sparse.msglmm(family, iterations, Y, X, M, offset, sigmab, sigmas, taub, beta0,
  Psi0, Sigma0, verbose = 1)
}
\arguments{
\item{X}{the design matrix.}

\item{A}{the adjacency matrix for the underlying graph. The matrix need not be binary, but it must be numeric and symmetric.}

\item{theta}{the vector of parameter values: \eqn{\theta = (\beta^\prime, \eta)^\prime}{\theta = (\beta', \eta)'}.}
}
\value{
A vector that is distributed exactly according to the centered autologistic model with the given design matrix and parameter values.
}
\description{
Fit a sparse multivariate SGLMM.
}
\details{
This function does things
}

