% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vcov.sparse.sglmm}
\alias{vcov.sparse.sglmm}
\title{Return the covariance matrix of the regression parameters of a \code{sparse.sglmm} model object.}
\usage{
\method{vcov}{sparse.sglmm}(object, ...)
}
\arguments{
\item{object}{a fitted \code{sparse.sglmm} model object.}

\item{\dots}{additional arguments.}
}
\value{
An estimate of the posterior covariance matrix of the regression coefficients.
}
\description{
Return the covariance matrix of the regression parameters of a \code{sparse.sglmm} model object.
}

